[codex][refactor] 拆分飞书卡片输出层
